+++ mpire/worker.py
@@ -2,13 +2,16 @@ import collections
 import queue
 import signal
 import traceback
+from datetime import datetime
 from functools import partial
+from multiprocessing.managers import ListProxy
 from typing import Any, Callable, Optional, Tuple
 
 import numpy as np
 
 from mpire.context import MP_CONTEXTS
 from mpire.exception import CannotPickleExceptionError, StopWorker
+from mpire.utils import TimeIt
 
 # If multiprocess is installed we want to use that as it has more capabilities than regular multiprocessing (e.g.,
 # pickling lambdas en functions located in __main__)
@@ -28,7 +31,9 @@ class AbstractWorker:
     def __init__(self, start_method: str, worker_id: int, tasks_queue: mp.JoinableQueue,
                  results_queue: mp.JoinableQueue, worker_done_array: mp.Array, task_completed_queue: mp.JoinableQueue,
                  exception_queue: mp.JoinableQueue, exception_lock: mp.Lock, exception_thrown: mp.Event,
-                 func_pointer: Callable, keep_order: bool, shared_objects: Any = None,
+                 func_pointer: Callable, keep_order: bool, start_time: datetime, worker_start_up_time: mp.Array,
+                 worker_n_completed_tasks: mp.Array, worker_waiting_time: mp.Array, worker_working_time: mp.Array,
+                 max_task_duration: mp.Array, max_task_args: ListProxy, shared_objects: Any = None,
                  worker_lifespan: Optional[int] = None, pass_worker_id: bool = False,
                  use_worker_state: bool = False) -> None:
         """
@@ -45,6 +50,14 @@ class AbstractWorker:
         :param func_pointer: Function pointer to call each time new task arguments become available
         :param keep_order: Boolean flag which signals if the task arguments contain an order index which should be
             preserved and not fed to the function pointer (e.g., used in ``map``)
+        :param start_time: `datetime` object indicating at what time the Worker instance was created and started
+        :param worker_start_up_time: Array object which holds the total number of seconds the workers take to start up
+        :param worker_n_completed_tasks: Array object which holds the total number of completed tasks per worker
+        :param worker_waiting_time: Array object which holds the total number of seconds the workers have been idle
+        :param worker_working_time: Array object which holds the total number of seconds the workers are executing the
+            task function
+        :param max_task_duration: Array object which holds the top 5 max task durations in seconds per worker
+        :param max_task_args: Array object which holds the top 5 task arguments (string) for the longest task per worker
         :param shared_objects: ``None`` or an iterable of process-aware shared objects (e.g., ``multiprocessing.Array``)
             to pass to the function as the first argument
         :param worker_lifespan: Number of chunks a worker can handle before it is restarted. If ``None``, workers will
@@ -66,6 +79,13 @@ class AbstractWorker:
         self.exception_thrown = exception_thrown
         self.func_pointer = func_pointer
         self.keep_order = keep_order
+        self.start_time = start_time
+        self.worker_start_up_time = worker_start_up_time
+        self.worker_n_completed_tasks = worker_n_completed_tasks
+        self.worker_waiting_time = worker_waiting_time
+        self.worker_working_time = worker_working_time
+        self.max_task_duration = max_task_duration
+        self.max_task_args = max_task_args
         self.shared_objects = shared_objects
         self.worker_lifespan = worker_lifespan
         self.pass_worker_id = pass_worker_id
@@ -73,6 +93,13 @@ class AbstractWorker:
 
         # Worker state
         self.worker_state = {}
+
+        # Additional worker insights container that holds (task duration, task args) tuples, sorted for heapq. We use a
+        # local container as to not put too big of a burden on interprocess communication
+        self.max_task_duration_list = (list(zip(self.max_task_duration[self.worker_id * 5:(self.worker_id + 1) * 5],
+                                                self.max_task_args[self.worker_id * 5:(self.worker_id + 1) * 5]))
+                                       if self.max_task_duration is not None else None)
+        self.max_task_duration_last_updated = datetime(1970, 1, 1)
 
         # Exception handling variables
         ctx = MP_CONTEXTS[self.start_method]
@@ -106,6 +133,10 @@ class AbstractWorker:
         Continuously asks the tasks queue for new task arguments. When not receiving a poisonous pill or when the max
         life span is not yet reached it will execute the new task and put the results in the results queue.
         """
+        # Store how long it took to start up
+        if self.worker_start_up_time is not None:
+            self.worker_start_up_time[self.worker_id] = (datetime.now() - self.start_time).total_seconds()
+
         # Obtain additional args to pass to the function
         additional_args = []
         if self.pass_worker_id:
@@ -124,10 +155,12 @@ class AbstractWorker:
         while self.worker_lifespan is None or n_chunks_executed < self.worker_lifespan:
 
             # Obtain new chunk of jobs
-            next_chunked_args = self._retrieve_task()
+            with TimeIt(self.worker_waiting_time, self.worker_id):
+                next_chunked_args = self._retrieve_task()
 
             # If we obtained a poison pill, we stop. When we receive None this means we stop because of an exception
             if next_chunked_args == '\0' or next_chunked_args is None:
+                self._update_task_insights()
                 if next_chunked_args == '\0':
                     self.tasks_queue.task_done()
                 return
@@ -141,7 +174,11 @@ class AbstractWorker:
                         # Try to run this function and save results
                         with self.is_running_lock:
                             self.is_running = True
-                        results.append(func(args))
+                        with TimeIt(self.worker_working_time, self.worker_id, self.max_task_duration_list,
+                                    lambda: self._format_args(args, separator=' | ')):
+                            results.append(func(args))
+                        if self.worker_n_completed_tasks is not None:
+                            self.worker_n_completed_tasks[self.worker_id] += 1
                         with self.is_running_lock:
                             self.is_running = False
 
@@ -171,7 +208,13 @@ class AbstractWorker:
             self.tasks_queue.task_done()
             n_chunks_executed += 1
 
+            # Update task insights every once in a while
+            if (self.max_task_duration is not None and
+                    (datetime.now() - self.max_task_duration_last_updated).total_seconds() > 2):
+                self._update_task_insights()
+
         # Notify WorkerPool to start a new worker if max lifespan is reached
+        self._update_task_insights()
         if self.worker_lifespan is not None and n_chunks_executed == self.worker_lifespan:
             self.worker_done_array[self.worker_id] = True
 
@@ -205,18 +248,9 @@ class AbstractWorker:
             # empty when the first one arrives.
             if not self.exception_thrown.is_set():
 
-                # Determine function arguments
-                func_args = args[1] if self.keep_order else args
-                if isinstance(func_args, dict):
-                    argument_str = "\n".join("Arg %s: %s" % (str(key), str(value)) for key, value in func_args.items())
-                elif isinstance(func_args, collections.abc.Iterable) and not isinstance(func_args, (str, bytes)):
-                    argument_str = "\n".join("Arg %d: %s" % (arg_nr, str(arg)) for arg_nr, arg in enumerate(func_args))
-                else:
-                    argument_str = "Arg 0: %s" % func_args
-
                 # Create traceback string
                 traceback_str = "\n\nException occurred in Worker-%d with the following arguments:\n%s\n%s" % (
-                    self.worker_id, argument_str, traceback.format_exc()
+                    self.worker_id, self._format_args(args), traceback.format_exc()
                 )
 
                 # Sometimes an exception cannot be pickled (i.e., we get the _pickle.PickleError: Can't pickle
@@ -231,6 +265,23 @@ class AbstractWorker:
                 # Add exception to queue
                 self.exception_queue.put((type(err), traceback_str))
                 self.exception_thrown.set()
+
+    def _format_args(self, args: Any, separator: str = '\n') -> str:
+        """
+        Format the function arguments to a string form.
+
+        :param args: Funtion arguments
+        :param separator: String to use as separator between arguments
+        :return: String containing the task arguments
+        """
+        # Determine function arguments
+        func_args = args[1] if self.keep_order else args
+        if isinstance(func_args, dict):
+            return separator.join("Arg %s: %s" % (str(key), repr(value)) for key, value in func_args.items())
+        elif isinstance(func_args, collections.abc.Iterable) and not isinstance(func_args, (str, bytes)):
+            return separator.join("Arg %d: %s" % (arg_nr, repr(arg)) for arg_nr, arg in enumerate(func_args))
+        else:
+            return "Arg 0: %s" % func_args
 
     def _helper_func_with_idx(self, func: Callable, args: Tuple[int, Any]) -> Tuple[int, Any]:
         """
@@ -268,6 +319,16 @@ class AbstractWorker:
             return func(*args)
         else:
             return func(args)
+
+    def _update_task_insights(self) -> None:
+        """
+        Update synced containers with new top 5 max task duration + args
+        """
+        if self.max_task_duration is not None:
+            task_durations, task_args = zip(*self.max_task_duration_list)
+            self.max_task_duration[self.worker_id * 5:(self.worker_id + 1) * 5] = task_durations
+            self.max_task_args[self.worker_id * 5:(self.worker_id + 1) * 5] = task_args
+            self.max_task_duration_last_updated = datetime.now()
 
 
 class ForkWorker(AbstractWorker, MP_CONTEXTS['fork'].Process):